+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - Emit event on producer mutation (create and update).
+- Emit event on consumer mutation (create and update).
 - Emit event on streambinding mutation.
 - Notification type and entity headers in Kafka messages